refactor: split CaptureWindowView into focused subviews (#22)
Extract CaptureSubredditPicker, CaptureLinksSection, and CaptureMediaSection
into their own files, reducing CaptureWindowView from 300 to 179 lines.
